Fixed issue when scrolling through pages in UI`PageViewController` in continuous mode causes an error loading the data from the API.
This call to refresh in the viewDidAppear was originally to ensure changes made on the Settings screen are reflected without the user having to manually pull to refresh. This is now redundant and is actually unnecessarily spamming the API, as the SettingsViewController now calls a closure on ViewDidDisappear, passing the responsibility of handling the refresh elsewhere.